typeDefs: create many files for each tables and call in the typeDefs file
