+++ src/typeDefs.ts
@@ -1,327 +1,23 @@
 import { gql } from 'apollo-server-express';
+import { drawsGql } from './typeDefs/draws';
+import { ordersGql } from './typeDefs/orders';
+import { ordersItemsGql } from './typeDefs/ordersItems';
+import { organizationsGql } from './typeDefs/organizations';
+import { shirtsGql } from './typeDefs/shirts';
+import { sizesGql } from './typeDefs/sizes';
+import { themesGql } from './typeDefs/themes';
+import { usersGql } from './typeDefs/users';
 
 const typeDefs = gql`
   scalar Date
-
-  type Users {
-    idUser: ID!
-    lastname: String
-    firstname: String
-    birthday: Date
-    phone: String
-    email: String!
-    password: String
-    address: String
-    zipCode: String
-    city: String
-    role: String
-    bio: String
-    orders: [Orders]
-    organizations: [Organizations]
-    draws: [Draws]
-  }
-
-  type Orders {
-    idOrder: ID!
-    price: Float
-    date: Date
-    address: String
-    zipCode: String
-    city: String
-    idUser: ID!
-    user: Users
-    shirts: [OrdersItems]
-  }
-
-  type Organizations {
-    idOrganization: ID!
-    name: String
-    phone: String
-    email: String
-    address: String
-    zipCode: String
-    city: String
-    idUser: ID!
-    user: Users
-    draws: [Draws]
-  }
-
-  type Themes {
-    idTheme: ID!
-    name: String
-    draws: [Draws]
-  }
-
-  type Draws {
-    idDraw: ID!
-    name: String
-    idUser: ID!
-    user: Users
-    idOrganization: ID!
-    organization: Organizations
-    idTheme: ID!
-    urlDraw: String
-    theme: Themes
-    sizes: [Shirts]
-  }
-
-  type Shirts {
-    idShirt: ID!
-    price: Float
-    idDraw: ID!
-    draw: Draws
-    idSize: ID!
-    size: Sizes
-    orders: [OrdersItems]
-  }
-
-  type Sizes {
-    idSize: ID!
-    name: String
-    draws: [Shirts]
-  }
-
-  type OrdersItems {
-    idOrderItem: ID!
-    quantity: Int
-    idShirt: ID!
-    shirt: Shirts
-    order: Orders
-  }
-
-  type Query {
-    AllUsers: [Users]
-    OneUser(idUser: ID!): Users
-    UserOrganization(idUser: ID!): [Organizations]
-    AllOrders: [Orders]
-    OneOrder(idOrder: ID!): Orders
-    AllOrganizations: [Organizations]
-    OneOrganization(idOrganization: ID!): Organizations
-    AllThemes: [Themes]
-    OneTheme(idTheme: ID!): Themes
-    AllDraws: [Draws]
-    OneDraw(idDraw: ID!): Draws
-    AllSizes: [Sizes]
-    OneSize(idSize: ID!): Sizes
-    AllShirts: [Shirts]
-    OneShirt(idShirt: ID!): Shirts
-    AllOrderItems: [OrdersItems]
-    OneOrderItem(idOrderItem: ID!): OrdersItems
-  }
-
-  input userCreateInput {
-    lastname: String
-    firstname: String
-    birthday: Date
-    phone: String
-    email: String!
-    password: String
-    address: String
-    zipCode: String
-    city: String
-    role: String
-    bio: String
-  }
-
-  input updateUserInput {
-    lastname: String
-    firstname: String
-    birthday: Date
-    phone: String
-    email: String!
-    password: String
-    address: String
-    zipCode: String
-    city: String
-    role: String
-    bio: String
-  }
-
-  type UpdateUserResponse {
-    message: String
-    user: Users
-  }
-
-  input createOrganizationInput {
-    idOrganization: Int
-    name: String
-    phone: String
-    email: String
-    address: String
-    zipCode: String
-    city: String
-    idUser: Int
-    siret: String
-  }
-
-  input updateOrganizationInput {
-    idOrganization: Int
-    name: String
-    phone: String
-    email: String
-    address: String
-    zipCode: String
-    city: String
-    idUser: Int
-    siret: String
-  }
-
-  type UpdateOrganizationResponse {
-    message: String
-    organization: Organizations
-  }
-
-  input createOrderInput {
-    idOrder: Int
-    price: Float
-    date: Date
-    address: String
-    zipCode: String
-    city: String
-    idUser: Int
-  }
-
-  input updateOrderInput {
-    idOrder: Int
-    price: Float
-    date: Date
-    address: String
-    zipCode: String
-    city: String
-    idUser: Int
-  }
-
-  type UpdateOrderResponse {
-    message: String
-    order: Orders
-  }
-
-  type DeleteOrganizationResponse {
-    message: String
-    organization: Organizations
-  }
-
-  input drawCreateInput {
-    idDraw: ID
-    name: String
-    idUser: Int!
-    idOrganization: Int
-    idTheme: Int!
-    urlDraw: String
-  }
-
-  input updateDrawInput {
-    idDraw: ID
-    name: String
-    idUser: Int!
-    idOrganization: Int
-    idTheme: Int!
-    urlDraw: String
-  }
-
-  type UpdateDrawResponse {
-    message: String
-    draw: Draws
-  }
-  type DeleteDrawResponse {
-    message: String
-    draw: Draws
-  }
-  input sizeCreateInput {
-    idSize: ID
-    name: String
-  }
-
-  input updateSizeInput {
-    idSize: ID
-    name: String
-  }
-
-  type UpdateSizeResponse {
-    message: String
-    size: Sizes
-  }
-  type DeleteSizeResponse {
-    message: String
-    size: Sizes
-  }
-
-  input themeCreateInput {
-    idTheme: ID
-    name: String
-  }
-
-  input updateThemeInput {
-    idTheme: ID
-    name: String
-  }
-
-  type UpdateThemeResponse {
-    message: String
-    theme: Themes
-  }
-  type DeleteThemeResponse {
-    message: String
-    theme: Themes
-  }
-
-  input orderItemCreateInput {
-    idOrderItem: ID
-    quantity: Int!
-    idShirt: Int!
-    idOrder: Int!
-  }
-
-  input updateOrderItemInput {
-    idOrderItem: ID
-    quantity: Int!
-    idShirt: Int!
-    idOrder: Int
-  }
-
-  type UpdateOrderItemResponse {
-    message: String
-    orderItem: OrdersItems
-  }
-  type DeleteOrderItemResponse {
-    message: String
-    orderItem: OrdersItems
-  }
-
-  type Mutation {
-    createUser(data: userCreateInput!): Users
-    updateUser(idUser: ID!, data: updateUserInput!): UpdateUserResponse
-    deleteUser(idUser: ID!): Users
-
-    createOrder(data: createOrderInput!): Orders
-    updateOrder(idOrder: ID!, data: updateOrderInput!): UpdateOrderResponse
-    deleteOrder(idOrder: ID!): Orders
-
-    createOrganization(data: createOrganizationInput!): Organizations
-    updateOrganization(
-      idOrganization: ID!
-      data: updateOrganizationInput!
-    ): UpdateOrganizationResponse
-    deleteOrganization(idOrganization: ID!): DeleteOrganizationResponse
-
-    createDraw(data: drawCreateInput!): Draws
-    updateDraw(idDraw: ID!, data: updateDrawInput!): UpdateDrawResponse
-    deleteDraw(idDraw: ID!): DeleteDrawResponse
-
-
-    createOrderItem(data: orderItemCreateInput!): OrdersItems
-    updateOrderItem(idOrderItem: ID!, data: updateOrderItemInput!): UpdateOrderItemResponse
-    deleteOrderItem(idOrderItem: ID!): DeleteOrderItemResponse
-
-    createSize(data: sizeCreateInput!): Sizes
-    updateSize(idSize: ID!, data: updateSizeInput!): UpdateSizeResponse
-    deleteSize(idSize: ID!): DeleteSizeResponse
-
-    createTheme(data: themeCreateInput!): Themes
-    updateTheme(idTheme: ID!, data: updateThemeInput!): UpdateThemeResponse
-    deleteTheme(idTheme: ID!): DeleteThemeResponse
-
-  }
+  ${usersGql}
+  ${ordersGql}
+  ${organizationsGql}
+  ${themesGql}
+  ${drawsGql}
+  ${sizesGql}
+  ${ordersItemsGql}
+  ${shirtsGql}
 `;
 
 export default typeDefs;
